GIW Count index files

+++ 02_GIW/02_Temario de Prácticas/01_Práctica Tema 1. Desarrollo de un sistema de recuperación de información mediante Lucene/Entrega de la práctica 1/p1_code/src/index/ExecuteIndex.java
@@ -11,7 +11,7 @@ public class ExecuteIndex {
         String dataDir = Constants.dataDir;
 
         Indexer indexer = new Indexer(indexDir);
-        int numIndexed;
+        Long numIndexed;
         long startTime = System.currentTimeMillis();
         numIndexed = indexer.createIndex(dataDir, new TextFileFilter());
         long endTime = System.currentTimeMillis();
